modules/requirements.txt
langchain
langchain-google-genai
faiss-cpu
PyPDF2
python-docx
unstructured
tiktoken
langchain-community
google.generativeai